clean up controls
